refactor player movement
