+++ src/game/classes/PlayerControls.ts
@@ -1,0 +1,55 @@
+import Player from "./Player";
+
+// PlayerControls.ts
+export default class PlayerControls {
+    private keyA!: Phaser.Input.Keyboard.Key;
+    private keyS!: Phaser.Input.Keyboard.Key;
+    private keyD!: Phaser.Input.Keyboard.Key;
+    private keyW!: Phaser.Input.Keyboard.Key;
+
+    private basePlayerSpeed = 300; // Move player speed to class level
+    private playerSpeed = this.basePlayerSpeed;
+
+    constructor(
+        private scene: Phaser.Scene,
+        private player: Player,
+    ) {
+        if (scene.input && scene.input.keyboard) {
+            this.keyA = scene.input.keyboard.addKey(
+                Phaser.Input.Keyboard.KeyCodes.A,
+            );
+            this.keyS = scene.input.keyboard.addKey(
+                Phaser.Input.Keyboard.KeyCodes.S,
+            );
+            this.keyD = scene.input.keyboard.addKey(
+                Phaser.Input.Keyboard.KeyCodes.D,
+            );
+            this.keyW = scene.input.keyboard.addKey(
+                Phaser.Input.Keyboard.KeyCodes.W,
+            );
+        }
+    }
+
+    update() {
+        // Player Movement
+        const velocity = { x: 0, y: 0 };
+
+        if (this.keyA.isDown) {
+            velocity.x = -this.playerSpeed;
+            this.player.anims.play("left", true);
+        } else if (this.keyD.isDown) {
+            velocity.x = this.playerSpeed;
+            this.player.anims.play("right", true);
+        }
+
+        if (this.keyW.isDown) {
+            velocity.y = -this.playerSpeed;
+        } else if (this.keyS.isDown) {
+            velocity.y = this.playerSpeed;
+            this.player.anims.play("turn");
+        }
+
+        this.player.setVelocity(velocity.x, velocity.y);
+    }
+}
+
